+++ anyblok/model/common.py
@@ -1,0 +1,30 @@
+# This file is a part of the AnyBlok project
+#
+#    Copyright (C) 2018 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v. 2.0. If a copy of the MPL was not distributed with this file,You can
+# obtain one at http://mozilla.org/MPL/2.0/.
+from anyblok.field import Field, FieldException
+
+MODEL = "MODEL"
+VIEW = "VIEW"
+
+
+def has_sql_fields(bases):
+    """ Tells whether the model as field or not
+
+    :param bases: list of Model's Class
+    :rtype: boolean
+    """
+    for base in bases:
+        for p in base.__dict__.keys():
+            try:
+                if hasattr(getattr(base, p), '__class__'):
+                    if Field in getattr(base, p).__class__.__mro__:
+                        return True
+            except FieldException:
+                # field function case already computed
+                return True
+
+    return False